Tests/LibWeb/Text/expected/display_list/blocking-wheel-listener-cache-replay.txt

  CompositorBlockingWheelEventRegion@1 rect=[8,8 100x100]
